basic/: audit chunk-1 follow-ups
* chaining_channels.py — use np.dtype(np.int32).itemsize for the
  bytes-to-elements conversion (the constant 4 was buried in a
  bytes->bytes*4->bytes//4 round-trip) and write n_elements_read as
  n_elements * 4 so the "4x larger compute buffer" relationship is
  visible at the call site.

* event_trace/test.py — fold two adjacent bo.read() rationale blocks
  into one sentence covering both symptoms (non-contiguous buffers and
  stale data) that motivate the bo.map() path.

* matrix_multiplication/single_core — Python --dtype_out default and
  README now both report i16 in / i32 out, matching the Makefile-driven
  defaults from makefile-common (the previous i16/i16 Python default
  diverged from `make run`).

* matrix_multiplication/whole_array — restore the --scalar flag and
  Makefile use_scalar=0 sentinel, plumbed through @iron.jit's
  Compile[bool] scalar param into kernels.mm(vectorized=not scalar).
  This brings back the debug knob for selecting matmul_scalar /
  zero_scalar variants on smaller sizes.  Smoke-built locally on both
  default (vector) and use_scalar=1 paths.

+++ programming_examples/basic/matrix_multiplication/single_core/single_core.py
@@ -213,7 +213,7 @@ def _make_argparser():
         "--dtype_out",
         type=str,
         choices=["bf16", "i8", "i16", "f32", "i32"],
-        default="i16",
+        default="i32",
     )
     p.add_argument("--b-col-maj", type=int, choices=[0, 1], default=0)
     p.add_argument(
